requirements.txt
PyYAML>=5.3.1
pandas>=1.1.0
yfinance>=0.1.70
nltk>=3.5
transformers>=4.0.0
torch>=1.7.0
scikit-learn>=0.24.0
matplotlib>=3.3.0
